redo torchscript example
